AgentWorkflow/doc_processor/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdf2image==1.17.0
python-pptx==1.0.2
Pillow==10.4.0
pdfminer.six==20231228
marker-pdf==0.2.17
